Add reweight module
